week_1_docker_terraform/requirements.txt
pandas==2.2.3
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10